requirements.txt
# MachineGPT - Python Dependencies
# Python 3.11+

# WEB FRAMEWORK
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5

# DATABASE
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
alembic==1.13.0

# AUTHENTICATION
PyJWT==2.8.0
bcrypt==4.1.2

# AI APIs
anthropic==0.34.0
voyageai==0.2.1
pinecone-client==3.0.0

# DOCUMENT PROCESSING
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23

# VIDEO/AUDIO PROCESSING
openai-whisper==20231117
moviepy==1.0.3
pydub==0.25.1
ffmpeg-python==0.2.0

# IMAGE PROCESSING
pytesseract==0.3.10
Pillow==10.1.0

# STORAGE
boto3==1.34.0

# UTILITIES
python-dotenv==1.0.0
requests==2.31.0
python-multipart==0.0.6

# VALIDATION
pydantic==2.5.3
email-validator==2.1.0

# TESTING
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0

# DEPLOYMENT
gunicorn==21.2.0
